backend/requirements.txt
fastapi
uvicorn
python-multipart
pandas
anthropic
python-dotenv
pypdf
